# docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Astrolabe API",
    "version": "0.3.0-beta.0",
    "description": "OpenClaw-first Responses router for OpenRouter with static manifests, virtual models, and safety policy."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    },
    {
      "url": "https://your-app-name.up.railway.app",
      "description": "Railway"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Healthy service response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/models": {
      "get": {
        "summary": "List Astrolabe models",
        "description": "Returns virtual Astrolabe models by default. Use `?view=raw` to inspect the static raw roster.",
        "responses": {
          "200": {
            "description": "Model list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/lanes": {
      "get": {
        "summary": "List lane manifests",
        "responses": {
          "200": {
            "description": "Lane list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/responses": {
      "post": {
        "summary": "Create a response",
        "description": "Primary Astrolabe endpoint. OpenClaw should use this endpoint with `api: openai-responses`.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponsesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful JSON or stream response",
            "headers": {
              "x-astrolabe-category": { "schema": { "type": "string" } },
              "x-astrolabe-lane": { "schema": { "type": "string" } },
              "x-astrolabe-initial-model": { "schema": { "type": "string" } },
              "x-astrolabe-final-model": { "schema": { "type": "string" } }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResource"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Confirmation or tool policy required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "Create a chat completion",
        "description": "Compatibility adapter over the same Astrolabe runtime used by `/v1/responses`.",
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful JSON or stream response",
            "headers": {
              "x-astrolabe-category": { "schema": { "type": "string" } },
              "x-astrolabe-lane": { "schema": { "type": "string" } },
              "x-astrolabe-initial-model": { "schema": { "type": "string" } },
              "x-astrolabe-final-model": { "schema": { "type": "string" } }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "service": { "type": "string" },
          "version": { "type": "string" },
          "routing_profile": { "type": "string" },
          "cost_efficiency_mode": { "type": "string" },
          "default_profile": { "type": "string" },
          "responses_enabled": { "type": "boolean" },
          "chat_completions_enabled": { "type": "boolean" }
        },
        "additionalProperties": true
      },
      "ListResponse": {
        "type": "object",
        "properties": {
          "object": { "type": "string" },
          "data": {
            "type": "array",
            "items": { "type": "object" }
          }
        },
        "additionalProperties": true
      },
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["messages"],
        "properties": {
          "model": { "type": "string" },
          "messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChatMessage" }
          },
          "stream": { "type": "boolean", "default": false },
          "tools": {
            "type": "array",
            "items": { "type": "object" }
          },
          "tool_choice": {},
          "response_format": {},
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "ResponsesRequest": {
        "type": "object",
        "required": ["input"],
        "properties": {
          "model": { "type": "string" },
          "input": {},
          "instructions": { "type": "string" },
          "stream": { "type": "boolean", "default": false },
          "tools": {
            "type": "array",
            "items": { "type": "object" }
          },
          "tool_choice": {},
          "response_format": {},
          "reasoning": {},
          "text": {},
          "max_output_tokens": { "type": "integer" },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      },
      "ChatMessage": {
        "type": "object",
        "properties": {
          "role": { "type": "string" },
          "content": {}
        },
        "required": ["role", "content"],
        "additionalProperties": true
      },
      "ChatCompletionResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "object": { "type": "string" },
          "created": { "type": "integer" },
          "choices": {
            "type": "array",
            "items": { "type": "object" }
          },
          "usage": { "type": "object" },
          "astrolabe": { "type": "object" }
        },
        "additionalProperties": true
      },
      "ResponseResource": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "object": { "type": "string" },
          "created": { "type": "integer" },
          "status": { "type": "string" },
          "output": {
            "type": "array",
            "items": { "type": "object" }
          },
          "usage": { "type": "object" },
          "metadata": { "type": "object" }
        },
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "type": { "type": "string" },
              "code": { "type": ["string", "null"] }
            },
            "additionalProperties": true
          }
        },
        "additionalProperties": true
      }
    }
  }
}